requirements.txt
numpy==1.24.2
opencv_python==4.7.0.72
Pillow==9.4.0
Pillow==9.5.0
PyAutoGUI==0.9.54
pyscreenshot==3.1
